pyproject.toml
[project]
name = "issue-triage-bot"
version = "1.0.0"
description = "Automatically categorize and route GitHub issues using Claude Code SDK"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "claude-code-sdk>=0.0.20",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.3.0",
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]

[tool.hatch.build.targets.wheel]
packages = ["."]
